create inverter.py, move code, add adrinverter fixture
